This commit updates cf-expandables: - Removes the ExpandableToggle.js file - Moves code from ExpandableToggle.js to Expandable.js - Updates atomic-component to v1.2.2
